+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Added operation `list-apps` to display current app installations for an organization. ([#101](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/101))
 - Added validation for secrets to not start with restricted prefix "GITHUB_". ([#100](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/100))
 - Added operation `dispatch-workflow` to dispatch a workflow in a specified repository.
 - Added flag `--update-filter` for plan, local-plan and apply operations to only update matching webhooks / secrets. ([#90](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/90))